#396: Introduce DeltaFactory.
