+++ src/treemodel/delta/weakinsertdelta.js
@@ -7,7 +7,7 @@
 
 import InsertDelta from './insertdelta.js';
 import { register } from '../batch.js';
-import { registerDeserializer } from './delta.js';
+import { registerDeserializer } from './deltafactory.js';
 import InsertOperation from '../operation/insertoperation.js';
 import NodeList from '../nodelist.js';
